fix(client): Use correct Mezmo api host, and consume v3 api responses
We have moved to a "v3" version of the APIs, and they have a response
body that contains `{meta, data}`.  This changes all URIs to be for v3,
and adds a generic struct for extracting the `data` portion of the
result. Also, point to the correct host that has the ingress rules for
publicly-routed pipeline APIs.

Ref: LOG-17553