chore(models): Add a base schemas and missing methods
In `pipeline-service`, we use a base schema to hold fields that are
present for all nodes such as `title, description, etc`. This commit
adds a similar pattern so we don't have to repeat common fields for
each added node type. Given the amount of types TF will support, it
minimizes effort in adding new ones, keeps code clean by highlighting
the unique properties, and mimics what pipeline service is doing in case
there are any changes to those base schemas down the road.

This commit also fills in the blank for missing methods that weren't
yet implemented. This allowed for a test to be added for sinks.

Ref: LOG-17765