fix(provider): Allow for errors to be created in `fromModelFunc`
The "from model" function is responsible for creating a component
structure suitable for transmission to CRUD apis, and each resource
implements this function. There will be cases where we need advanced
validation based on what's in the plan versus what's in the state. The
function has access to those things and can generate errors if need be.

Change the function signature to return the `Component` as well as a
`Diagnostics` structure.  This is the preferred TF way to handle errors
since the implementor can choose to make an error or warning, leaving
the caller to simply check whether errors have been made. Those errors
and/or warnings are also aggregated, letting the framework handle them
if they ever bubble up.

Ref: LOG-17791