chore(test): Add `GetCachedConfig` to minimize repeating code
The way the TF provider tests work, the `Config` has to be provided **in
full** for every test. On more complicated components and pipelines,
this will become a lot of repetetive code. This commit adds functions
that use an in-memory map to save and retrieve configuration text
blocks so the developer doesn't have to repeat so much TF configuration
code.

Ref: LOG-17826