fix: Out-of-sync state is causing false positive test assertions
The way that state is managed in Terraform is that it expects models to
mirror api responses, and when an api response is received, the state is
completely overwritten with that response. Our modeling does NOT match
the api results directly since we break out "user_config" fields for a
variety of reasons. Because of this, we need to manually update
the plan/state fields with the response.  However, if we fail to update
any fields, the original state values still exist (as populated by the
user-inputted terraform config). This causes our tests to *think* that
values in the state have also been created on the remote resource, but
it's not necessarily true.

This commit adds a function that will nullify all of our user config
fields prior to calling Create, Read or Update methods so that if we
fail to update them with the api's response, it will error. All we need
to do is use Go tags to denote which struct fields belong to
`user_config:"true"`.

The remainder of the changes in this commit are bug fixes for things
that started erroring after the fix was in place.

Ref: LOG-18104