refactor(tests): Clean up component tests to follow pattern
A number of the test cases for the pipeline components did not follow
the pattern used by the remainder. Each test case that needs to track
state should use a unique cache key for itself and that should be aware
that the same type of component could appear as both a source and
destination.

Ref: LOG-18405