fix: Ensure processors can be converted to terraform model
ConvertToTerraformModel populates list fields as empty and
object fields as null due to the absence of a Terraform state.
Models that retrieve the list item type or object type from
the model object then fail.
To fix this, we need to check if the type is nil and retrieve
it directly from the model's schema.

Ref: LOG-18449