chore(docs): Changes for `go generate`
Running `go generate` needed a sum update for the uuid package in ordre
to work right. Beyond that, running `go generate` now removes trailing
empty lines from most/all of the files. This is most likely due to a
bump in a dependency, possibly with the recent bump of the TF framework.

Ref: LOG-18373