fix: Update parse and sequential parse proccessors to accept new options
New regex options were added to the API which broke CI with
parse and sequential parse processors failing in tests.
Four of the 5 new options have been added to the regex parser.
A new function has also been added to exclude options we
do not yet know about in Terraform.

Ref: LOG-18741