feat(alerts): Support pipline alerts (v2)
This commit sets up support for using pipeline alerts and its schema.
The schema has several aspects of mutual exclusion or inclusion which
the UI handles gracefully. In Terraform, we don't have such luxury, so
we'll ask the use to be explicit about values (not many "defaults" are
used), and we will manually check for some exclusive values enforced by
the schema.

This commit also re-works our `conditional` schemas such that the
logical operators can be specified. This is needed because change alerts
will only allow for "value change" logical operators.

Ref: LOG-19605
Ref: LOG-19606
Ref: LOG-19607