refactor(provider)!: Change `gateway_route_id` to `shared_source_id`
This provider should not expose internal terminology such as what
"gateway routes" are used for. To a user, these are shared sources, and
they should be renamed as such. This also fits in with the recent
addition of creating shared sources independently, since they are
already called "shared sources."

BREAKING CHANGE: The renaming of this field in the schema will break
users of this provider, and thus this should be a `major` release.

Ref: LOG-19968