feat(provider): Add `publish_pipeline` resource
This resource can monitor a pipeline for changes and automatically
publish it. To manage flow control such that Terraform will only publish
a pipeline after all of its components are created/updated, the use of
child modules is imperative when using this resource. This way, a
`depends_on` can be used to ensure the publish runs last.

Ref: LOG-18021