fix(client): Parse all JSON responses, including errors
Previously, the provider would spit out the serialized JSON as part of
the error. Not only is this impossible to read, but it prevents the
provider from introspecting the errors to decide if it needs to care
about them. This commit adds a formal error structure and marshaling for
errors and `code`. Certain validation errors we're choosing to skip
because they don't make sense to the user and are in addition to
whatever the root cause is.

Ref: LOG-20073