fix(client): Parse all JSON responses, including errors
Previously, the provider would spit out the serialized JSON as part of
the error. Not only is this impossible to read, but it prevents the
provider from introspecting the errors to decide if it needs to care
about them. This commit adds a formal error structure and marshaling for
errors and `code`. Certain validation errors we're choosing to skip
because they don't make sense to the user and are in addition to
whatever the root cause is.

Another part of this change is a refactor for `readBody`. It became
confusing to have 2 things trying and read the body stream because
that's a destructive operation to the stream. This commit combines
`readJson` and `readBody` into one function that will read the body AND
decode its json if it's not empty.

Ref: LOG-20073