fix(resources)!: `publish_pipeline` should be re-created every time
Selectively publishing a pipeline doesn't really work in Terraform due
to race conditions. The pipeline resources are updated atomically
outside of the scope of `publish_pipeline`, therefore it's possible that
states will update even if the publish command fails (paused pipeline,
for example). This causes a situation where no changes are detected, and
thus `publish_pipeline` does not become part of the `plan`.

To fix this, constantly remove the `publish_pipeline` from state every
time `Read()` is called. This will force "re-creation" of the publish
resource such that it will be executed every time. Since the
`ENOCHANGES` error is ignored, this operation becomes idempotent.

BREAKING CHANGE: This removes the required `updated_at` property from
the schema. Users will get an error if `updated_at` is required, so this
is a breaking change.

Ref: LOG-20085