test: docs and split unit and acceptance tests
Follow convention by prefixing every acceptance test with TestAcc and
divide the test target into two, one for units and another for
acceptance tests. This makes it easier to run just the unit tests which
complete way faster.

Also add a docs test so we can ensure that the documentation is valid
before merging. The docs diff here is fixing a few broken examples.

Ref: LOG-19481