refactor(client): Combine api responses with provider logging
When executing the provider, `fmt.Println` commands are ignored.
Conversely, `tflog` commands are ignored when running tests locally.
This changes the logging so that it will print for both based on the
single env var used for provider logging, `TF_LOG_PROVIDER=TRACE`.

Using `TRACE` will allow a developer to see API
requests/responses/errors when running tests or executing the provider.

Ref: LOG-20134